Complete

+++ SalesTaxManager/Configuration/TrolleyContent.cs
@@ -1,16 +1,12 @@
 ﻿using SalesTaxManager.Entities;
 using SalesTaxManager.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace SalesTaxManager.Configuration
 {
     public class ItemContent
     {
         public string Product { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
         public bool  Imported { get; set; }
     }
